Add files via upload
This is a contact tracing code that allows for different data to be pulled from the original list of people that have been in contact with each other in the txt file DataSet1. In this txt file the first name on every line is seen to be the one who is in contact with the others in the line of text.